# docs/swagger.yaml
basePath: /api/
definitions:
...
  models.ErrorResponse:
    properties:
      message:
        example: An error occurred
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
...
  models.SuccessResponseCreateUUID:
    properties:
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      message:
        example: Object created successfully
        type: string
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        example: "2024-01-01T10:00:00Z"
        type: string
      email:
        example: johndoe@example.com
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      is_active:
        example: true
        type: boolean
      last_login:
        example: "2024-12-01T15:30:00Z"
        type: string
      name:
        example: John
        type: string
      role_id:
        example: admin
        type: string
      surname:
        example: Doe
        type: string
      username:
        example: johndoe
        type: string
    type: object
...
  models.UsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/models.UserResponse'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: Simple API for managing ticket reservations.
  title: Ticket Reservation API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Retrieve a list of all users, including their details and roles.
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            $ref: '#/definitions/models.UsersResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users (admin only)
      tags:
      - users
    put:
      description: Retrieve a user, including its details and roles.
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/models.SuccessResponseCreateUUID'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new user.
      tags:
      - users
  /users/{id}:
    delete:
      description: Deletes user from the database (only owner/admin).
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user.
      tags:
      - users
    get:
      description: Retrieve a user, including its details and roles.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/models.UserResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a user by ID (admin only)
      tags:
      - users
    put:
      description: Update user details (only owner/admin).
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user.
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"